Initial working release This is a sight reading app which you can use to learn how to read sheet music. It features a toy piano on screen as well as midi support (in chrome at least) for your real digital piano.
This is an MVP of sorts which I'm putting into source control now
that I seem to have a working version.
I will clean up the code in the near future as well as perhaps
add a few other game modes and improvements.
I'm thinking of making it at least display random chrods if not
random melodies, or, why not, actual songs.